git-svn-id: http://svn.codehaus.org/geoserver/branches/1.4.x@5301 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 685404db8742d63aeef63f27125d2df281a24f1f [formerly 0fb709c17bd2321abc8849d9cc6c9166ec0a1849]
Former-commit-id: d8d49a527cff981b0d234d63827167833267ac79

+++ web/src/main/java/org/vfny/geoserver/action/MapPreviewAction.java
@@ -95,9 +95,9 @@ public class MapPreviewAction extends GeoServerAction
 		
 		// 2) delete any existing generated files in the generation directory
 		ServletContext sc = request.getSession().getServletContext();
-		File rootDir =  GeoserverDataDirectory.getGeoserverDataDirectory(sc);
-		File previewDir = new File(sc.getRealPath("data/generated"));
-		
+		//File rootDir =  GeoserverDataDirectory.getGeoserverDataDirectory(sc);
+		//File previewDir = new File(sc.getRealPath("data/generated"));
+		File previewDir = new File(sc.getRealPath("preview"));
 		//File previewDir = new File(rootDir, "data/generated");
 		if (!previewDir.exists())
 			previewDir.mkdirs();
@@ -173,8 +173,9 @@ public class MapPreviewAction extends GeoServerAction
 			else
 			{
 				if (f.exists())
-					if (!f.delete())	// try to delete the file
-						throw new FileNotFoundException("could not delete file: "+f.getName());
+					f.delete();
+//					if (!f.delete())	// try to delete the file
+//						throw new FileNotFoundException("could not delete file: "+f.getName());
 			}
 		}
 		
@@ -213,23 +214,33 @@ public class MapPreviewAction extends GeoServerAction
 		File config_file = new File(previewDir, ft_namespace + "_"+ ft_name+"Config.xml");
 		File xml_file = new File(previewDir, ft_namespace + "_"+ ft_name+".xml");
 		
+		//JD: making the exists check, since they dont have to be generated
+		// every time
+		
 		// *.html
-		FileOutputStream html_fos = new FileOutputStream(html_file);
-		PrintStream html_out = new PrintStream(html_fos);
-		createIndexHTML(html_out, ft_name, ft_namespace);
-		html_out.close();
+		if ( ! html_file.exists() ) {
+			FileOutputStream html_fos = new FileOutputStream(html_file);
+			PrintStream html_out = new PrintStream(html_fos);
+			createIndexHTML(html_out, ft_name, ft_namespace);
+			html_out.close();	
+		}
 		
-		// *Config.xml
-		FileOutputStream config_fos = new FileOutputStream(config_file);
-		PrintStream config_out = new PrintStream(config_fos);
-		createConfigXML(config_out, ft_name, ft_namespace);
-		config_out.close();
+		if ( ! config_file.exists() ) {
+			//*Config.xml
+			FileOutputStream config_fos = new FileOutputStream(config_file);
+			PrintStream config_out = new PrintStream(config_fos);
+			createConfigXML(config_out, ft_name, ft_namespace);
+			config_out.close();	
+		}
 		
-		// *.xml
-		FileOutputStream xml_fos = new FileOutputStream(xml_file);
-		PrintStream xml_out = new PrintStream(xml_fos);
-		createLayersXML(xml_out, ft_name, ft_namespace, bbox);
-		xml_out.close();
+		if ( !xml_file.exists() ) {
+			//*.xml
+			FileOutputStream xml_fos = new FileOutputStream(xml_file);
+			PrintStream xml_out = new PrintStream(xml_fos);
+			createLayersXML(xml_out, ft_name, ft_namespace, bbox);
+			xml_out.close();	
+		}
+		
 	}
 	
 	
@@ -250,7 +261,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("");
 		out.println("<link rel=\"stylesheet\" href=\"../../style.css\"");
 		out.println("	type=\"text/css\">");
-		out.println("<link rel=\"stylesheet\" href=\"../mbdemos/lib/skin/default/html.css\"");
+		out.println("<link rel=\"stylesheet\" href=\"../mb/lib/skin/default/html.css\"");
 		out.println("	type=\"text/css\">");
 		out.println("");
 		out.println("<script type=\"text/javascript\">");
@@ -258,7 +269,7 @@ public class MapPreviewAction extends GeoServerAction
 		// CHANGE HERE
 		out.println("      var mbConfigUrl='"+ft_namespace+"_"+ft_name+"Config.xml"+"';");
 		out.println("    </script>");
-		out.println("<script type=\"text/javascript\" src=\"../mbdemos/lib/Mapbuilder.js\"></script>");
+		out.println("<script type=\"text/javascript\" src=\"../mb/lib/Mapbuilder.js\"></script>");
 		out.println("</head>");
 		out.println("");
 		out.println("<body onload=\"mbDoLoad()\">");
@@ -287,7 +298,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("        <td align=\"left\" id=\"mapScaleText\"/>");
 		out.println("				<td align=\"right\">Powered by <a href=\"http://mapbuilder.sourceforge.net\">Community Map");
 		out.println("				Builder</a></td>");
-		out.println("				<td><img src=\"../mbdemos/lib/skin/default/images/Icon.gif\" alt=\"\" /></td>");
+		out.println("				<td><img src=\"../mb/lib/skin/default/images/Icon.gif\" alt=\"\" /></td>");
 		out.println("			</tr>");
 		out.println("		</table>");
 		out.println("		</td>");
@@ -343,7 +354,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("        </MapPane>");
 		out.println("        <AoiBoxWZ id=\"aoiBox2\">");
 		out.println("          <htmlTagId>mainMapPane</htmlTagId>");
-		out.println("          <stylesheet>../mbdemos/lib/widget/GmlRenderer.xsl</stylesheet>");
+		out.println("          <stylesheet>../mb/lib/widget/GmlRenderer.xsl</stylesheet>");
 		out.println("          <lineColor>#FF0000</lineColor>");
 		out.println("          <lineWidth>1</lineWidth>");
 		out.println("          <crossSize>15</crossSize>");
@@ -378,7 +389,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("        </MapPane>");
 		out.println("        <AoiBoxWZ id=\"aoiBox2\">");
 		out.println("          <htmlTagId>locatorMap</htmlTagId>");
-		out.println("          <stylesheet>../mbdemos/lib/widget/GmlRenderer.xsl</stylesheet>");
+		out.println("          <stylesheet>../mb/lib/widget/GmlRenderer.xsl</stylesheet>");
 		out.println("          <lineColor>#FF0000</lineColor>");
 		out.println("          <lineWidth>1</lineWidth>");
 		out.println("          <crossSize>15</crossSize>");
@@ -456,7 +467,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("      <tooltip xml:lang=\"fr\">redonner la carte ses dimensions compl???ts</tooltip>");
 		out.println("    </Reset>");
 		out.println("  </widgets>");
-		out.println("  <skinDir>../mbdemos/lib/skin/default</skinDir>");
+		out.println("  <skinDir>../mb/lib/skin/default</skinDir>");
 		out.println("</MapbuilderConfig>");
 
 
@@ -486,7 +497,7 @@ public class MapPreviewAction extends GeoServerAction
 		out.println("    <Layer queryable=\"1\" hidden=\"0\">");
 		// CHANGE HERE
 		out.println("      <Server service=\"OGC:WMS\" version=\"1.1.1\" title=\""+ft_namespace+":"+ft_name+" Preview\">");
-		out.println("        <OnlineResource xlink:type=\"simple\" xlink:href=\"../../wms\"/>");
+		out.println("        <OnlineResource xlink:type=\"simple\" xlink:href=\"../wms\"/>");
 		out.println("      </Server>");
 		// CHANGE HERE
 		out.println("      <Name>"+ft_namespace+":"+ft_name+"</Name>");
